Use ValueRequest in factory function. Make ValueRequest implementation detail
